automod: mad science, attempting to factor engine/event/effect apart.
Non-compiling checkpoint.

Separating events and effects feels good.  Read-only event structures
and append-only effects structures clarifies many semantics.

All of the persistence logic itself stayed close to the engine, and
takes the effects struct as a parameter.  This felt super good.
It means all the methods on the effects struct aire simply for
mutating it.  The struct itself remains legible and has no side-effects
on its own (except logging), so the ability to test this without
entanglements should be quite high.

Distinct packages for events and effects is up for debate.  It made it
easier to have the compiler help me enforce detanglement during the
experiment, but is not necessarily a good idea in the long run.

The biggest issue that needs further pondering is what we want to see
the ruletypes file look like.  We aimed to take one more step from here
to introduce a "business context" object and reduce the number of
parameters of rule funcs to approximately one (maybe two, for unpacked
specific values like "post")... but the trouble with that idea is that
we still see that e.g. IdentityEvent and RecordEvent are still already
distinct types.  We note that RecordEvent is a strict superset of
IdentityEvent, so maybe we just decide to take the superset one in
order to have a single simple "business context" object... but we're
not yet entirely sure if that flies the way we like.
