+++ automod/engine/engine_test.go
@@ -1,4 +1,4 @@
-package automod
+package engine
 
 import (
 	"context"
