+++ automod/engine/fetch.go
@@ -1,4 +1,4 @@
-package automod
+package engine
 
 import (
 	"context"
@@ -74,40 +74,4 @@ func (e *Engine) FetchAndProcessRecent(ctx context.Context, atid syntax.AtIdenti
 		}
 	}
 	return nil
-}
-
-type AccountCapture struct {
-	CapturedAt  syntax.Datetime                     `json:"capturedAt"`
-	AccountMeta AccountMeta                         `json:"accountMeta"`
-	PostRecords []comatproto.RepoListRecords_Record `json:"postRecords"`
-}
-
-func (e *Engine) CaptureRecent(ctx context.Context, atid syntax.AtIdentifier, limit int) (*AccountCapture, error) {
-	ident, records, err := e.FetchRecent(ctx, atid, limit)
-	if err != nil {
-		return nil, err
-	}
-	pr := []comatproto.RepoListRecords_Record{}
-	for _, r := range records {
-		if r != nil {
-			pr = append(pr, *r)
-		}
-	}
-
-	// clear any pre-parsed key, which would fail to marshal as JSON
-	ident.ParsedPublicKey = nil
-	am, err := e.GetAccountMeta(ctx, ident)
-	if err != nil {
-		return nil, err
-	}
-
-	// auto-clear sensitive PII (eg, account email)
-	am.Private = nil
-
-	ac := AccountCapture{
-		CapturedAt:  syntax.DatetimeNow(),
-		AccountMeta: *am,
-		PostRecords: pr,
-	}
-	return &ac, nil
 }
