+++ automod/engine/ruleset.go
@@ -1,9 +1,11 @@
-package automod
+package engine
 
 import (
 	"fmt"
 
 	appbsky "github.com/bluesky-social/indigo/api/bsky"
+	"github.com/bluesky-social/indigo/automod/effects"
+	"github.com/bluesky-social/indigo/automod/event"
 )
 
 type RuleSet struct {
@@ -14,15 +16,12 @@ type RuleSet struct {
 	IdentityRules     []IdentityRuleFunc
 }
 
-func (r *RuleSet) CallRecordRules(evt *RecordEvent) error {
+func (r *RuleSet) CallRecordRules(evt *event.RecordEvent, eff *effects.RecordEffect) error {
 	// first the generic rules
 	for _, f := range r.RecordRules {
-		err := f(evt)
+		err := f(evt, eff)
 		if err != nil {
 			return err
-		}
-		if evt.Err != nil {
-			return evt.Err
 		}
 	}
 	// then any record-type-specific rules
@@ -33,12 +32,9 @@ func (r *RuleSet) CallRecordRules(evt *RecordEvent) error {
 			return fmt.Errorf("mismatch between collection (%s) and type", evt.Collection)
 		}
 		for _, f := range r.PostRules {
-			err := f(evt, post)
+			err := f(evt, eff, post)
 			if err != nil {
 				return err
-			}
-			if evt.Err != nil {
-				return evt.Err
 			}
 		}
 	case "app.bsky.actor.profile":
@@ -47,39 +43,30 @@ func (r *RuleSet) CallRecordRules(evt *RecordEvent) error {
 			return fmt.Errorf("mismatch between collection (%s) and type", evt.Collection)
 		}
 		for _, f := range r.ProfileRules {
-			err := f(evt, profile)
+			err := f(evt, eff, profile)
 			if err != nil {
 				return err
 			}
-			if evt.Err != nil {
-				return evt.Err
-			}
 		}
 	}
 	return nil
 }
 
-func (r *RuleSet) CallRecordDeleteRules(evt *RecordDeleteEvent) error {
+func (r *RuleSet) CallRecordDeleteRules(evt *event.RecordDeleteEvent, eff *effects.RecordDeleteEffect) error {
 	for _, f := range r.RecordDeleteRules {
-		err := f(evt)
+		err := f(evt, eff)
 		if err != nil {
 			return err
-		}
-		if evt.Err != nil {
-			return evt.Err
 		}
 	}
 	return nil
 }
 
-func (r *RuleSet) CallIdentityRules(evt *IdentityEvent) error {
+func (r *RuleSet) CallIdentityRules(evt *event.IdentityEvent, eff *effects.IdentityEffect) error {
 	for _, f := range r.IdentityRules {
-		err := f(evt)
+		err := f(evt, eff)
 		if err != nil {
 			return err
-		}
-		if evt.Err != nil {
-			return evt.Err
 		}
 	}
 	return nil
